Simplify severity pointer casts and move MatchesPath tests to pathutil.
