+++ pkg/engine/inspector_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/open-policy-agent/opa/rego"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/DataDog/datadog-iac-scanner/internal/pathutil"
 	"github.com/DataDog/datadog-iac-scanner/internal/tracker"
 	"github.com/DataDog/datadog-iac-scanner/pkg/detector"
 	"github.com/DataDog/datadog-iac-scanner/pkg/engine/source"
@@ -1019,38 +1018,6 @@ func TestInspector_checkComment(t *testing.T) {
 			if got := checkComment(tt.line, tt.lines); got != tt.want {
 				t.Errorf("checkComment() = %v, want %v", got, tt.want)
 			}
-		})
-	}
-}
-
-func TestMatchesPath(t *testing.T) {
-	tests := []struct {
-		name    string
-		pattern string
-		file    string
-		want    bool
-	}{
-		{"exact match", "/repo/src/main.tf", "/repo/src/main.tf", true},
-		{"directory prefix", "/repo/src", "/repo/src/main.tf", true},
-		{"directory prefix no trailing sep", "/repo/src", "/repo/src/nested/main.tf", true},
-		{"no match different dir", "/repo/test", "/repo/src/main.tf", false},
-		{"no partial prefix match", "/repo/s", "/repo/src/main.tf", false},
-		{"glob star", "/repo/src/*.tf", "/repo/src/main.tf", true},
-		{"glob star no match", "/repo/src/*.tf", "/repo/src/main.go", false},
-		{"glob question mark", "/repo/src/mai?.tf", "/repo/src/main.tf", true},
-		// double-star (**) glob
-		{"doublestar all tf files", "/repo/**/*.tf", "/repo/src/main.tf", true},
-		{"doublestar nested", "/repo/**/*.tf", "/repo/src/nested/deep/main.tf", true},
-		{"doublestar no match ext", "/repo/**/*.tf", "/repo/src/main.go", false},
-		{"doublestar prefix any subdir", "**/terraform/**", "infra/terraform/main.tf", true},
-		{"doublestar prefix no match", "**/terraform/**", "infra/k8s/main.yaml", false},
-		{"doublestar matches zero segments", "a/**/b.tf", "a/b.tf", true},
-		{"doublestar matches one segment", "a/**/b.tf", "a/x/b.tf", true},
-		{"doublestar matches many segments", "a/**/b.tf", "a/x/y/z/b.tf", true},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, pathutil.MatchesPath(tt.pattern, tt.file))
 		})
 	}
 }
